+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ Features:
 
 * Add ``APISpec.add_tags`` method for adding Swagger tags. Thanks :user:`martinlatrille`.
 
+Bug fixes:
+
+* Fix bug in marshmallow extension where metadata was being lost when converting marshmallow ``Schemas`` when ``many=False``. Thanks again :user:`martinlatrille`.
+
 0.8.0 (2016-03-06)
 ++++++++++++++++++